requirements.txt
asgiref==3.8.1
Babel==2.15.0
config==0.5.1
Django==4.1.13
django-cors-headers==4.3.1
django-money==3.5.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
psycopg2-binary==2.9.9
py-moneyed==3.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.2.4
typing_extensions==4.11.0
uritemplate==4.1.1
